Move Protocol to rtc_ice_candidate_stats module
