+++ crates/native/src/api/mod.rs
@@ -16,7 +16,7 @@ use libwebrtc_sys as sys;
 
 pub use self::{
     rtc_ice_candidate_stats::{
-        CandidateType, IceCandidateStats, RtcIceCandidateStats,
+        CandidateType, IceCandidateStats, Protocol, RtcIceCandidateStats,
     },
 };
 // Re-exporting since it is used in the generated code.
@@ -367,30 +367,6 @@ impl From<sys::RTCStatsIceCandidatePairState>
             sys::RTCStatsIceCandidatePairState::kFailed => Self::Failed,
             sys::RTCStatsIceCandidatePairState::kSucceeded => Self::Succeeded,
             _ => unreachable!(),
-        }
-    }
-}
-
-/// Transport protocols used in [WebRTC].
-///
-/// [WebRTC]: https://w3.org/TR/webrtc
-pub enum Protocol {
-    /// [Transmission Control Protocol][1].
-    ///
-    /// [1]: https://en.wikipedia.org/wiki/Transmission_Control_Protocol
-    Tcp,
-
-    /// [User Datagram Protocol][1].
-    ///
-    /// [1]: https://en.wikipedia.org/wiki/User_Datagram_Protocol
-    Udp,
-}
-
-impl From<sys::Protocol> for Protocol {
-    fn from(protocol: sys::Protocol) -> Self {
-        match protocol {
-            sys::Protocol::Tcp => Self::Tcp,
-            sys::Protocol::Udp => Self::Udp,
         }
     }
 }
